refactor: extract shared test result types from ghost-stories
