+++ code/core/src/shared/utils/test-result-types.ts
@@ -1,0 +1,31 @@
+export interface StoryTestResult {
+  storyId: string;
+  status: 'PASS' | 'FAIL' | 'PENDING';
+  error?: string;
+  stack?: string;
+  /** Whether the story rendered to an empty/invisible DOM element */
+  emptyRender?: boolean;
+}
+
+export interface CategorizedError {
+  category: string;
+  count: number;
+  uniqueCount: number;
+  matchedDependencies: string[];
+}
+
+export interface ErrorCategorizationResult {
+  totalErrors: number;
+  categorizedErrors: Record<string, CategorizedError>;
+  uniqueErrorCount: number;
+}
+
+export interface TestRunAnalysis {
+  total: number;
+  passed: number;
+  passedButEmptyRender: number;
+  successRate: number;
+  successRateWithoutEmptyRender: number;
+  uniqueErrorCount: number;
+  categorizedErrors: Record<string, CategorizedError>;
+}
